main.dart
Added ChangeNotifierProvider for managing state with PetsProvider.
Set up GoRouter for routing between home and add page.
Defined routes for HomePage and AddPage.
DioClient class

Implemented getPets method to fetch pets from the API using Dio.
Implemented createPet method to add a new pet to the API using FormData.
Handled DioError exceptions and error logging in both methods.
PetsProvider class

Added a list of pets and initialized it with a default pet.
Implemented getPets method to fetch pets from the API.
Added getPetsWithNotify method to fetch pets and notify listeners.
Implemented createPet method to add a new pet and update the UI by notifying listeners.
AddForm widget

Created a form for adding a pet with fields for name, age, gender, and image selection.
Integrated form validation and submission logic.
Added functionality to pick an image from the gallery.
Used the Provider to submit form data and create a new pet.
AddPage widget

Created a scaffold for the add page with an AppBar.
Embedded the AddForm widget as the body of the page.